代码格式化

+++ src/main/java/org/b3log/solo/util/MD5.java
@@ -15,12 +15,11 @@
  */
 package org.b3log.solo.util;
 
-
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 
 /**
  * MD5 hash.
@@ -30,67 +29,69 @@ import org.slf4j.LoggerFactory;
  */
 public final class MD5 {
 
-    /**
-     * Logger.
-     */
-    private static Logger logger = LoggerFactory.getLogger(MD5.class);
+	/**
+	 * Logger.
+	 */
+	private static Logger logger = LoggerFactory.getLogger(MD5.class);
 
-    /**
-     * Message digest.
-     */
-    private static MessageDigest messageDigest;
+	/**
+	 * Message digest.
+	 */
+	private static MessageDigest messageDigest;
 
-    /**
-     * Low 8 bits all sets to 1.
-     */
-    private static final int LOW_8_BITS_1 = 0xff;
+	/**
+	 * Low 8 bits all sets to 1.
+	 */
+	private static final int LOW_8_BITS_1 = 0xff;
 
-    /**
-     * Append size.
-     */
-    private static final int APPEND_SIZE = 16;
+	/**
+	 * Append size.
+	 */
+	private static final int APPEND_SIZE = 16;
 
-    /**
-     * Private default constructor.
-     */
-    private MD5() {}
+	/**
+	 * Private default constructor.
+	 */
+	private MD5() {
+	}
 
-    static {
-        try {
-            messageDigest = MessageDigest.getInstance("MD5");
-        } catch (final NoSuchAlgorithmException e) {
-            logger.error(e.getMessage(), e);
-            throw new RuntimeException(e);
-        }
-    }
+	static {
+		try {
+			messageDigest = MessageDigest.getInstance("MD5");
+		} catch (final NoSuchAlgorithmException e) {
+			logger.error(e.getMessage(), e);
+			throw new RuntimeException(e);
+		}
+	}
 
-    /**
-     * Hashs(MD5) the specified string.
-     *
-     * @param string the specified string
-     * @return hashed string from the specified string
-     */
-    public static String hash(final String string) {
-        final char[] charArray = string.toCharArray();
-        final byte[] byteArray = new byte[charArray.length];
+	/**
+	 * Hashs(MD5) the specified string.
+	 *
+	 * @param string
+	 *            the specified string
+	 * @return hashed string from the specified string
+	 */
+	public static String hash(final String string) {
+		final char[] charArray = string.toCharArray();
+		final byte[] byteArray = new byte[charArray.length];
 
-        for (int i = 0; i < charArray.length; i++) {
-            byteArray[i] = (byte) charArray[i];
-        }
+		for (int i = 0; i < charArray.length; i++) {
+			byteArray[i] = (byte) charArray[i];
+		}
 
-        final byte[] bytes = messageDigest.digest(byteArray);
-        final StringBuffer hexValue = new StringBuffer();
+		final byte[] bytes = messageDigest.digest(byteArray);
+		final StringBuffer hexValue = new StringBuffer();
 
-        for (int i = 0; i < bytes.length; i++) {
-            final int val = ((int) bytes[i]) & LOW_8_BITS_1;
+		for (int i = 0; i < bytes.length; i++) {
+			final int val = (bytes[i]) & LOW_8_BITS_1;
 
-            if (val < APPEND_SIZE) {
-                hexValue.append("0");
-            }
+			if (val < APPEND_SIZE) {
+				hexValue.append("0");
+			}
 
-            hexValue.append(Integer.toHexString(val));
-        }
+			hexValue.append(Integer.toHexString(val));
+		}
 
-        return hexValue.toString();
-    }
+		return hexValue.toString();
+	}
 }
